refactor: move input components into the inputs folder.
