+++ src/app/(auth)/_components/forms/SignInForm.tsx
@@ -7,8 +7,8 @@ import { useForm } from 'react-hook-form';
 import { useDispatch } from 'react-redux';
 import SubmitButton from '@/app/(auth)/_components/SubmitButton';
 import ErrorMessage from '@/components/shared/ErrorMessage';
-import FormInput from '@/components/shared/FormInput';
-import PasswordInput from '@/components/shared/PasswordInput';
+import FormInput from '@/components/shared/inputs/FormInput';
+import PasswordInput from '@/components/shared/inputs/PasswordInput';
 import { Form, FormField } from '@/components/ui/form';
 import { signIn } from '@/lib/actions/auth/sign-in';
 import { setUserAndStatus } from '@/lib/store/slices/user-slice';
